requirements.txt
asgiref==3.11.1
colorama==0.4.6
Django==5.2.11
gunicorn==23.0.0
django-filter==25.2
djangorestframework==3.16.1
iniconfig==2.3.0
packaging==26.0
pluggy==1.6.0
psycopg2-binary==2.9.11
Pygments==2.19.2
pytest==9.0.2
pytest-django==4.12.0
sqlparse==0.5.5
tzdata==2025.3
Pillow
requests>=2.28.0
